+++ SplunkAppForWazuh/appserver/static/js/controllers/management/rules/managerRulesetCtrl.js
@@ -19,7 +19,8 @@ define(['../../module', './ruleset'], function (controllers, Ruleset) {
       $tableFilterService,
       $csvRequestService,
       $restartService,
-      isAdmin
+      isAdmin,
+      $fileEditor
     ) {
       super(
         $scope,
@@ -29,7 +30,8 @@ define(['../../module', './ruleset'], function (controllers, Ruleset) {
         $currentDataService,
         $tableFilterService,
         $csvRequestService,
-        $restartService
+        $restartService,
+        $fileEditor
       )
       this.isAdmin = isAdmin
     }
@@ -44,8 +46,6 @@ define(['../../module', './ruleset'], function (controllers, Ruleset) {
       this.scope.$broadcast('wazuhSearch', { term: '', removeFilters: true })
       this.scope.addNewFile = () => this.addNewFile()
       this.scope.saveRuleConfig = (fileName, dir, overwrite) => this.saveRuleConfig(fileName, dir, overwrite)
-      this.scope.closeEditingFile = () => this.closeEditingFile()
-      this.scope.xmlIsValid = valid => this.xmlIsValid(valid)
 
       this.scope.selectedNavTab = 'rules'
 
@@ -69,31 +69,12 @@ define(['../../module', './ruleset'], function (controllers, Ruleset) {
      */
     addNewFile() {
       this.scope.overwrite = false
-      this.scope.addingNewFile = true
       this.scope.editingFile = {
         file: ``,
         dir: `rules`
       }
       this.scope.addingNewFile = true
       this.scope.fetchedXML = `<!-- Configure your local rules here -->`
-    }
-
-    /**
-     * Edit rules and decoders functions
-     */
-    closeEditingFile() {
-      this.scope.editingFile = false
-      this.scope.addingNewFile = false
-      this.scope.fetchedXML = ''
-    }
-
-    /**
-     * Check if XML is valid
-     * @param {Boolean} valid 
-     */
-    xmlIsValid(valid) {
-      this.scope.xmlHasErrors = valid
-      this.scope.$applyAsync()
     }
 
     /**